rename ios tests
